+++ sdk/metric/internal/x/features.go
@@ -3,7 +3,10 @@
 
 package x // import "go.opentelemetry.io/otel/sdk/metric/internal/x"
 
-import "strconv"
+import (
+	"strconv"
+	"strings"
+)
 
 // MetricExportBatchSize is an experimental feature flag that controls the
 // max export batch size for metric data.
@@ -18,5 +21,20 @@ var MetricExportBatchSize = newFeature(
 			return val, true
 		}
 		return 0, false
+	},
+)
+
+// ParallelCallbacks is an experimental feature flag that enables running
+// observable-instrument callbacks concurrently during a collection.
+//
+// To enable this feature set the OTEL_GO_X_PARALLEL_CALLBACKS environment
+// variable to the case-insensitive string value of "true".
+var ParallelCallbacks = newFeature(
+	[]string{"PARALLEL_CALLBACKS"},
+	func(v string) (bool, bool) {
+		if strings.EqualFold(v, "true") {
+			return true, true
+		}
+		return false, false
 	},
 )